рассчёт стоимости заказа вынесен из редьюсера: помещён в компонент конструктора бургеров и мемоизирован; исправлен баг с потерей булок при сортировке начинки
